basic functionality is ready to go

+++ index.js
@@ -1,38 +1,45 @@
 'use strict';
 
-const { parsePackageJson, generateInstallDevString } = require('./lib/packageJsonParser')  ;
-const { writeFile, removeOldFile, appendFile } = require('./lib/utils');
+const {parsePackageJson, generateInstallString, generateInstallDevString} = require('./lib/package-json-parser');
+const {writeFile, appendFile} = require('./lib/utils');
 
 const path = require('path');
-const fs = require('fs');
 const currentPath = process.cwd();
 
-async function entry(input, {postfix = 'rainbows'} = {}) {
-	let forceMode = input.f === true;
-	let outputFile = input.output || false;
-	let interactiveMode = input.i === true;
-	let reportMode = input.report === true;
+async function entry(input) {
+	const forceMode = input.f === true;
+	const outputFile = input.output || false;
+	const interactiveMode = input.i === true;
+	const reportMode = input.report === true;
 
 	const fullPath = path.join(currentPath, 'package.json');
 
 	const installs = await parsePackageJson(fullPath);
-	const dependencies = generateInstallDevString(installs.dependencies);
+	const dependencies = generateInstallString(installs.dependencies);
 	const devDependencies = generateInstallDevString(installs.devDependencies);
 
-	let reportString = "";
-
-	if(reportMode) {
-		console.log("oida");
+	if (forceMode) {
+		return ':) Force mode not implemented yet';
 	}
 
-	if(outputFile) {
+	if (outputFile) {
 		const trimmedFilePath = outputFile.trim();
-		// await removeOldFile(trimmedFilePath);
-		await writeFile(trimmedFilePath, dependencies)
-		await appendFile(trimmedFilePath, devDependencies)
+		// Await removeOldFile(trimmedFilePath);
+		await writeFile(trimmedFilePath, dependencies);
+		await appendFile(trimmedFilePath, devDependencies);
+
+		return `:) You can find your strings inside '${outputFile}'`;
 	}
 
-	return `${input} & ${postfix}`;
+	if (interactiveMode) {
+		return ':) Interactive mode not implemented yet';
+	}
+
+	if (reportMode) {
+		return ':) Report mode not implemented yet';
+	}
+
+	return `${dependencies}\n${devDependencies}`;
 }
 
 module.exports = entry;
